Allow large IDs in the database

+++ athena/athena/models/db_feedback.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Boolean, String, Float, JSON, UniqueConstraint
+from sqlalchemy import Column, BigInteger, Boolean, String, Float, JSON, UniqueConstraint
 
 from .model import Model
 
@@ -6,12 +6,12 @@ from .model import Model
 class DBFeedback(Model):
     __table_args__ = (UniqueConstraint('lms_id'),)
 
-    id = Column(Integer, primary_key=True, index=True, autoincrement=True)
-    lms_id = Column(Integer)
+    id = Column(BigInteger, primary_key=True, index=True, autoincrement=True)
+    lms_id = Column(BigInteger)
     title = Column(String)
     description = Column(String)
     credits = Column(Float, nullable=False)
-    grading_instruction_id = Column(Integer)
+    grading_instruction_id = Column(BigInteger)
     meta = Column(JSON, nullable=False)
 
     # not in the schema, but used in the database to distinguish between feedbacks and feedback suggestions
