+++ athena/athena/models/db_text_feedback.py
@@ -1,6 +1,6 @@
 from typing import Optional
 
-from sqlalchemy import Column, Integer, ForeignKey
+from sqlalchemy import Column, Integer, BigInteger, ForeignKey
 from sqlalchemy.orm import relationship
 
 from athena.database import Base
@@ -13,8 +13,8 @@ class DBTextFeedback(DBFeedback, Base):
     index_start: Optional[int] = Column(Integer)  # type: ignore
     index_end: Optional[int] = Column(Integer)  # type: ignore
 
-    exercise_id = Column(Integer, ForeignKey("text_exercises.id", ondelete="CASCADE"), index=True)
-    submission_id = Column(Integer, ForeignKey("text_submissions.id", ondelete="CASCADE"), index=True)
+    exercise_id = Column(BigInteger, ForeignKey("text_exercises.id", ondelete="CASCADE"), index=True)
+    submission_id = Column(BigInteger, ForeignKey("text_submissions.id", ondelete="CASCADE"), index=True)
 
     exercise = relationship("DBTextExercise", back_populates="feedbacks")
     submission = relationship("DBTextSubmission", back_populates="feedbacks")
